requirements.txt
celery==4.3.0
pytest==5.1.2
hbmqtt==0.9.5
python-decouple==3.1
requests==2.21.0